backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.999
requests==2.31.0
pymongo==4.6.1
torch==2.9.0
numpy==2.1.2
pydantic==2.8.0
websockets==12.0
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
